Implemented SPN algorithm. Adjusted decimal points in the output. Refactored the inherited classes constructors to use super(). Did half of the documentation. Still a few classes to go.

+++ src/Process.java
@@ -3,21 +3,21 @@
 //Programmer: Liam Craft - c3339847
 //Date: 02/09/2020
 
-import java.util.ArrayList;
-
 public class Process {
 	private String id;
 	private int arrive;			//arrival time into the ready queue
-	private float start;			//actual start time of the process
-	private float finish;			//actual finishing time
+	private int start;			//actual start time of the process
+	private int finish;			//actual finishing time
 	private int execSize;		//Time for this process to execute fully
 	private int execCount;		//Used when a process gets interrupted to track how long it has been executing.
 	private int priority;		//The execution priority given to this process
-	private float tATime;		//turn around time: actual time to complete the job
-	private float waitTime;		//The time this process spent waiting, (total time - arrival time) - execSize
+	private int tATime;		//turn around time: actual time to complete the job
+	private int waitTime;		//The time this process spent waiting, (total time - arrival time) - execSize
 	private int quantum;		//Time slice given to this process in time switching algorithms like Round Robin.
 	
 	//default
+	//Precondition: None
+	//Postconditions:
 	public Process() {
 		this.id = "";
 		this.arrive = 0;
@@ -31,16 +31,18 @@ public class Process {
 		this.quantum = 0;
 	}
 	//constructor
+	//Precondition: None
+	//Postconditions:
 	public Process(String newId, int newArrive, int newExecSize, int newPriority) {
 		this.id = newId;
 		this.arrive = newArrive;
 		this.execSize = newExecSize;
 		this.priority = newPriority;
-		this.start = 0f;
-		this.finish = 0f;
+		this.start = 0;
+		this.finish = 0;
 		this.execCount = 0;
-		this.tATime = 0f;
-		this.waitTime = 0f;
+		this.tATime = 0;
+		this.waitTime = 0;
 		this.quantum = 0;
 	}
 	
@@ -55,11 +57,11 @@ public class Process {
 		this.arrive = newArrive;
 	}
 	//set the start time for this process
-	public void setStart(float newStart) {
+	public void setStart(int newStart) {
 		this.start = newStart;
 	}
 	//set the finishing time for this process
-	public void setFinish(float newFinish) {
+	public void setFinish(int newFinish) {
 		this.finish = newFinish;
 	}
 		
@@ -77,12 +79,12 @@ public class Process {
 	}
 	
 	// Set the calculated Turn around time for this process
-	public void setTATime(float newTATime) {
+	public void setTATime(int newTATime) {
 		this.tATime = newTATime;
 	}
 	
 	//Set the total time this process spent in a waiting queue
-	public void setWaitTime(float newWaitTime) {
+	public void setWaitTime(int newWaitTime) {
 		this.waitTime = newWaitTime;
 	}
 	
@@ -106,10 +108,10 @@ public class Process {
 	public int getArrive() {
 		return this.arrive;
 	}
-	public float getStart() {
+	public int getStart() {
 		return this.start;
 	}
-	public float getFinish() {
+	public int getFinish() {
 		return this.finish;
 	}
 	public int getExecSize() {
@@ -121,10 +123,10 @@ public class Process {
 	public int getPriority() {
 		return this.priority;
 	}
-	public float getTATime() {
+	public int getTATime() {
 		return this.tATime;
 	}
-	public float getWaitTime() {
+	public int getWaitTime() {
 		return this.waitTime;
 	}
 	public int getQuantum() {
